Fix focus and mark across the docs
