+++ code_snippets/kotlin/src/main/kotlin/usecases/workflows/WorkflowSubmitter.kt
@@ -7,20 +7,20 @@ class WorkflowSubmitter {
     // <start_here>
     // import dev.restate.sdk.client.Client;
     val rs = Client.connect("http://localhost:8080")
-    // mark
+    // !mark
     SignupWorkflowClient.fromClient(rs, user.id)
-        // mark
+        // !mark
         .submit(user)
 
     // do something else, with workflow running in the background
 
     // attach back to the workflow
-    // mark
+    // !mark
     val result =
         SignupWorkflowClient.fromClient(rs, user.id)
-            // mark
+            // !mark
             .workflowHandle()
-            // mark
+            // !mark
             .attach()
     // <end_here>
   }
