+++ code_snippets/python/src/concepts/invocations/delayed.py
@@ -10,6 +10,6 @@ my_service = Service("MyService")
 # <start_delayed_call>
 @my_service.handler()
 async def my_handler(ctx: Context, arg):
-    # focus
+    # !focus
     ctx.service_send(greet, arg="Hi", send_delay=timedelta(seconds=1))
 # <end_delayed_call>
