+++ code_snippets/python/src/concepts/invocations/one_way.py
@@ -8,6 +8,6 @@ my_service = Service("MyService")
 # <start_one_way_call>
 @my_service.handler()
 async def my_handler(ctx: Context, arg):
-    # focus
+    # !focus
     ctx.service_send(greet, arg="Hi")
 # <end_one_way_call>
