+++ code_snippets/python/src/concepts/invocations/rpc.py
@@ -8,6 +8,6 @@ my_service = Service("MyService")
 # <start_rpc_call>
 @my_service.handler()
 async def my_handler(ctx: Context, arg):
-    # focus
+    # !focus
     greeting = await ctx.service_call(greet, arg="Hi")
 # <end_rpc_call>
