+++ code_snippets/ts/src/concepts/invocations/delayed.ts
@@ -3,11 +3,11 @@ import { greetCounter, greeter } from "./utils";
 
 // <start_delayed_call>
 async function myRestateHandler(ctx: restate.Context) {
-  // focus
+  // !focus
   ctx
-    // focus
+    // !focus
     .serviceSendClient(greeter)
-    // focus
+    // !focus
     .greet({ greeting: "Hi" }, restate.rpc.sendOpts({ delay: 1000 }));
 }
 // <end_delayed_call>
