+++ code_snippets/ts/src/develop/clients/ingress.ts
@@ -61,13 +61,13 @@ const servicesAttach = async () => {
   // Send a message
   const handle = await rs
     .serviceSendClient(greeterService)
-    // mark
+    // !mark
     .greet(request, clients.rpc.sendOpts({ idempotencyKey: "abcde" }));
 
   // ... do something else ...
 
   // Attach later to retrieve the result
-  // mark
+  // !mark
   const response = await rs.result(handle);
   // <end_service_attach>
 };
