+++ docs/concepts/invocations.mdx
@@ -8,7 +8,7 @@ import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import FeatureList from "../../src/components/FeatureList"
 import {Scrollycoding} from "../../src/components/code/scrollycoding";
-import {Code, CodeWithTabs} from "../../src/components/code/code";
+import {CodeWithTabs} from "../../src/components/code/code";
 
 # Invocations
 
@@ -248,7 +248,7 @@ Duplicate requests will get the same response as the first request, or will latc
 
 ```bash
 curl localhost:8080/GreeterService/greet \
-# focus
+# !focus
   -H 'idempotency-key: ad5472esg4dsg525dssdfa5loi' \
   -H 'content-type: application/json' \
   -d '"Hi"'
