+++ docs/develop/java/clients.mdx
@@ -59,7 +59,7 @@ Invoke a handler programmatically with the SDK clients as follows:
         <TabItem value="kotlin" label="Kotlin">
             <TextAndCode>
                 **Request-response invocations** allow you to wait on a response from the handler.
-
+    
                 ```kotlin !result
                 CODE_LOAD::kotlin/src/main/kotlin/develop/clients/Ingress.kt#rpc_kotlin
                 ```
