+++ docs/operate/invocation.mdx
@@ -34,7 +34,7 @@ restate invocations list
     <summary>Output</summary>
 
     ```shell
-    # focus[36:74]
+    # !focus
     ❯ [2024-04-16 15:28:20.237 +02:00] inv_1aiqX0vFEFNH0TF1pLRFBDFosQCCTAN1M5 [CartObject @ Mary]::addTicket
     Status:      backing-off  (4 minutes, 42 seconds and 653 ms. Retried 32 time(s). Next
     retry in in 9 seconds and 469 ms))
