+++ docs/operate/versioning.mdx
@@ -28,11 +28,8 @@ To check which deployments are currently serving your services, do the following
 
             ```shell
             NAME             REVISION  FLAVOR  DEPLOYMENT TYPE  DEPLOYMENT ID
-            # mark[57:82]
             🌎  TicketObject     1                 HTTP 2           dp_14LsPzGz9HBxXIeBoH5wYUh
-            # mark[57:82]
             🌎  CartObject       1                 HTTP 2           dp_14LsPzGz9HBxXIeBoH5wYUh
-            # mark[57:82]
             🌎  CheckoutService  1                 HTTP 2           dp_14LsPzGz9HBxXIeBoH5wYUh
             ```
 
@@ -133,7 +130,7 @@ To update it, we deploy a new revision at `http://greeter-v2/`, and then registe
                 "services": [
             {
                 "name": "Greeter",
-                // focus
+                // !focus
                 "revision": 2
             }
                 ]
