+++ src/css/codehike-styling.css
@@ -94,3 +94,15 @@ ul.tabs {
     margin-block-start: 0;
     margin-block-end: 0;
 }
+
+.focus {
+    opacity: .5;
+}
+
+.focus[data-focus="true"] {
+    opacity: 1;
+}
+
+.no-border {
+    border: none;
+}
