+++ src/plugins/code-loader.js
@@ -109,7 +109,7 @@ const plugin = (options) => {
                     }
                     finalLines.push(line);
                 }
-                // console.log(JSON.stringify(finalLines))
+                console.log(JSON.stringify(finalLines))
             });
         } else {
             finalLines = lines;
